refactor(ios): split slash bubble out of EngineMessageRow.swift
EngineMessageRow.swift grew from 770 to 919 lines this branch despite
its @file-size-exception annotation. The annotation acknowledged the
file as a god file pending Workstream C decomposition but the slash-
bubble feature added this branch could and should live separately.

Extract userBubbleContentWithSlash, parseSlashCommand, and the
SlashCommandSegments/slashCommandPattern parser into a new sibling
file EngineMessageRow+SlashBubble.swift. The split is purely
organizational: call sites in conversationUserBubble and
engineUserBubble invoke the extension method by name as before.

Widens @Environment(\.appTheme) theme from private to internal so the
extension in a separate file can reference it — same module, no API
surface change.

Drops EngineMessageRow.swift from 919 to 854 lines. The file-size
exception annotation is updated to record the decomposition.
